requirements.txt
# Venice Bar Analyzer - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP Clients
requests==2.31.0
httpx==0.26.0

# Data Processing
pydantic==2.5.3
python-dotenv==1.0.0

# Google Places & Populartimes
googlemaps==4.10.0
populartimes @ git+https://github.com/m-wrz/populartimes.git

# NLP (for review analysis)
nltk==3.8.1
textblob==0.17.1

# ML (optional, for photo analysis)
# opencv-python==4.9.0.0
# tensorflow==2.15.0

# Utilities
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
